+++ src/main/java/org/ays/emergency_application/port/adapter/EmergencyEvacuationApplicationAdapter.java
@@ -1,4 +1,4 @@
-package org.ays.emergency_application.port.impl;
+package org.ays.emergency_application.port.adapter;
 
 import lombok.RequiredArgsConstructor;
 import org.ays.common.model.AysPage;
